Changed M104 to M109 to avoid hotend stalls
When the temperature is increased with M109, the hotend stays at the same spot which causes in-burn. It's better to gradually increase the temp using M104, which doesn't wait until the hotend reaches the target temperature.
This is probably the reason why most temperature towers go from hot to cool - the hotend can cool faster than it heats usually.